- Filter livraisons and demandes by status in LivraisonTable and DemandeTable components - Update Dashboard component to use "Demande Livraison" instead of "Livraisons" in subItems - Update login functionality to store user ID in local storage - Add API functions for managing demandes de livraisons - Remove unused Google Maps script from index.html and dynamically load it in index.js - Update AddressAutocomplete.js to handle missing postal code in address components - Added redux and finished logic for codePostal from address - Added Calendar version for readonly to render later in the dashboard

+++ src/components/clients/Clients.js
@@ -5,6 +5,9 @@ import Dashboard from "../dashboard/Dashboard";
 import Search from "../searchbar/Search";
 import ClientForm from "./ClientForm";
 import ClientTable from "./ClientTable";
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+
 const Clients = () => {
   const [clients, setClients] = useState([]);
   const [filteredClients, setFilteredClients] = useState([]);
@@ -14,15 +17,19 @@ const Clients = () => {
     first_name: '',
     last_name: '',
     address1: '',
+    code_postal: '',
     address2: '',
+    code_postal2: '',
     phone: ''
   });
   const [isEditMode, setIsEditMode] = useState(false);
   const [currentClient, setCurrentClient] = useState(null);
   const [isSearchActive, setIsSearchActive] = useState(false);
+
   useEffect(() => {
     fetchClientsData();
   }, [currentPage]);
+
   const fetchClientsData = async () => {
     try {
       const data = await fetchClients(currentPage);
@@ -32,30 +39,40 @@ const Clients = () => {
       console.error('Error fetching clients', error);
     }
   };
+
   const handleDelete = async (id) => {
-    try {
-      await deleteClient(id);
-      fetchClientsData();
-    } catch (error) {
-      console.error('Error deleting client', error);
+    if (window.confirm('Êtes-vous sûr de vouloir supprimer ce client?')) {
+      try {
+        await deleteClient(id);
+        fetchClientsData();
+        toast.success('Client supprimé avec succès!');
+      } catch (error) {
+        console.error('Error deleting client', error);
+        toast.error('Erreur lors de la suppression du client.');
+      }
     }
   };
+
   const handleModify = (client) => {
     setCurrentClient(client);
     setNewClient(client);
     setIsEditMode(true);
     setShowForm(true);
   };
+
   const handleEditClient = async (e) => {
     e.preventDefault();
     try {
       await modifyClient(newClient);
       fetchClientsData();
       setShowForm(false);
+      toast.success('Client modifié avec succès!');
     } catch (error) {
       console.error('Error modifying client', error);
+      toast.error('Erreur lors de la modification du client.');
     }
   };
+
   const handleSearch = async (searchTerm) => {
     if (searchTerm) {
       setIsSearchActive(true);
@@ -66,34 +83,39 @@ const Clients = () => {
       setFilteredClients(clients);
     }
   };
+
   const handleAddClient = async (e) => {
     e.preventDefault();
     try {
       await addClient(newClient);
       fetchClientsData();
       setShowForm(false);
+      toast.success('Client ajouté avec succès!');
     } catch (error) {
       console.error('Error adding client', error);
+      toast.error('Erreur lors de l\'ajout du client.');
     }
   };
+
   const handlePageChange = useCallback((page) => {
     setCurrentPage(page);
   }, []);
-  const handleChange=useCallback((e)=>{
-    const{name}=e.target;
-    setNewClient(prevState=>({
+
+  const handleChange = useCallback((e) => {
+    const { name, value } = e.target;
+    setNewClient(prevState => ({
       ...prevState,
-      [name]:e.target.value
-    }))
+      [name]: value
+    }));
+  }, []);
 
-
-  },[])
   return (
     <div className="flex">
       <Dashboard />
-      <div className="flex-1 container mx-auto p-9 relative mt-20 ">
+      <div className="flex-1 container mx-auto p-9 relative mt-20">
+        <ToastContainer />
         <Search setData={handleSearch} title={"Tous les clients"} />
-              <button
+        <button
           className="custom-color2 text-white px-4 py-2 rounded mb-4 absolute top-0 right-0 mt-4 mr-4 shadow hover:bg-blue-600 transition"
           onClick={() => {
             setShowForm(true);
@@ -102,7 +124,9 @@ const Clients = () => {
               first_name: '',
               last_name: '',
               address1: '',
+              code_postal: '',
               address2: '',
+              code_postal2: '',
               phone: ''
             });
           }}
@@ -111,18 +135,19 @@ const Clients = () => {
         </button>
         {showForm && (
           <ClientForm
-          newClient={newClient}
-          handleChange={handleChange}
-          handleAddClient={handleAddClient}
-          handleEditClient={handleEditClient}
-          setShowForm={setShowForm}
-          isEditMode={isEditMode}
+            newClient={newClient}
+            handleChange={handleChange}
+            handleAddClient={handleAddClient}
+            handleEditClient={handleEditClient}
+            setShowForm={setShowForm}
+            isEditMode={isEditMode}
           />
         )}
         <ClientTable clients={isSearchActive ? filteredClients : clients} handleDelete={handleDelete} handleModify={handleModify} />
         {!isSearchActive && <Pagination currentPage={currentPage} setCurrentPage={setCurrentPage} />}
-        </div>
+      </div>
     </div>
-  )
+  );
 }
+
 export default Clients;
